% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binned_cnvs.R
\name{binned_cnvs}
\alias{binned_cnvs}
\title{Compute the overlap between a CNV table and a list of genomic bins}
\usage{
binned_cnvs(
  cnvs,
  format = c("long", "count", "wide", "both"),
  bins = binned_genome(chrs_sten = QCtreeCNV::hg19_start_end_centromeres, bin_size =
    1e+05, chrs = 1:22)
)
}
\arguments{
\item{cnvs}{usual CNV `data.table`}

\item{format}{format for the output table, "count" and "both" return a list}

\item{bins}{for computing the overlap, by default are regular bins created by the helper function `binned_genome`}
}
\description{
This can be regular windows throughout the genome or exons / genes for example.
}
