% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cnv.R
\name{check_cnv}
\alias{check_cnv}
\title{Plot LRR/BAF for a given CNV}
\usage{
check_cnv(
  cnv,
  samp,
  snps = NULL,
  in_out_ratio = 1,
  adjusted_lrr = T,
  min_lrr = -1.2,
  max_lrr = 1,
  shrink_lrr = NULL
)
}
\arguments{
\item{cnv}{see load_snps_tbx() documentation}

\item{samp}{see load_snps_tbx() documentation}

\item{snps}{see load_snps_tbx() documentation}

\item{in_out_ratio}{see load_snps_tbx() documentation}

\item{adjusted_lrr}{see load_snps_tbx() documentation}

\item{min_lrr}{see load_snps_tbx() documentation}

\item{max_lrr}{see load_snps_tbx() documentation}

\item{shrink_lrr}{see load_snps_tbx() documentation}
}
\description{
This function is mostly needed for developing to check how CNVs looks like
"normally" and compare with the PNG form.
}
