% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCNVRs_IOU.R
\name{cnvrs_iou}
\alias{cnvrs_iou}
\title{Create CNVRs in a set of CNVS based on network analysis}
\usage{
cnvrs_iou(
  cnvs,
  chr_arm,
  screen_size = 500,
  min_iou = 0.75,
  leiden_res = 1,
  plots_path = NA,
  min_n = 20,
  max_force_merge_rounds = 5,
  force_merge_min_overlap = 0.75,
  arms_save_plots = c("2p", "16p", "15q", "19p")
)
}
\arguments{
\item{cnvs}{usual CNVs `data.table`}

\item{screen_size}{size of the comb that screen the cnvs to look for individual networks}

\item{min_iou}{minimum IOU filter. Define how similar two CNVs must be in order to be considered connected}

\item{leiden_res}{lower value will tend to create more smaller communities}

\item{min_n}{min n for recomputing CNVRs (small CNVRs might "drown" in very large networks)}

\item{max_force_merge_rounds}{how many rounds of force merge are performed at the end of the function}

\item{chr_arms}{chromsome arms location, from `QCtreeCNV` package}

\item{plot_path}{not yet implemented}
}
\description{
A network is defined a set of CNVs that overlap with each other. Usually there
are several ones in each chromosomal arm. Given a minimum IOU value
each pairs of CNVs is considered as two connected nodes. The CNVRs are
computed using community the detection algorithm 'leiden'
}
