% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tabix.R
\name{load_snps_tbx}
\alias{load_snps_tbx}
\title{Load intensity data in R from tabix file}
\usage{
load_snps_tbx(cnv, samp, snps = NULL, adjusted_lrr = T)
}
\arguments{
\item{cnv}{one line data.table in the usual cnv format}

\item{samp}{one line samples file in the usual format}

\item{snps}{the snps file for PennCNV, as data.table}

\item{adjusted_lrr}{load GC-adjusted LRR on normal LRR (T/F)}

\item{shrink_lrr}{}
}
\description{
A function to lead the snps data from the tabix indexed intensity file.
}
