% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{make_predicitons}
\alias{make_predicitons}
\title{Make Prediction on a set of CNVs}
\usage{
make_predicitons(model, root, cnvs)
}
\arguments{
\item{model}{pre trained model loaded using `luz::luz_load()`}

\item{root}{root folder for the dataset, created using `save_pngs_prediction()`}

\item{cnvs}{cnv data.table in the usual format}
}
\description{
This function uses a pre trained model to make a prediction on a new set
of CNVs
}
