% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{make_predictions}
\alias{make_predictions}
\title{Make Prediction on a set of CNVs}
\usage{
make_predictions(model, root, cnvs, return_pred_dt = F, batches = 1000)
}
\arguments{
\item{model}{pre trained model loaded using `luz::luz_load()`}

\item{root}{root folder for the dataset, created using `save_pngs_prediction()`}

\item{cnvs}{cnv data.table in the usual format}

\item{return_pred_dt}{returt the prediction table before merging w/ the cnv table, set
it to T if there is a problem with the `sample_ID`}
}
\description{
This function uses a pre trained model to make a prediction on a new set
of CNVs
}
