% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iou_matrix.R
\name{plot_iou}
\alias{plot_iou}
\title{IOU heatmap-like plot}
\usage{
plot_iou(dt)
}
\arguments{
\item{dt}{output of `iou_matrix()`, one chr arm at the time (one list item)}
}
\description{
IOU heatmap-like plot
}
