% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_association.R
\name{process_quick_res}
\alias{process_quick_res}
\title{Process results from quick assoc}
\usage{
process_quick_res(
  quick_res,
  markers,
  padj = F,
  max_pval = 0.1,
  loci,
  loci_boundaries = 250000
)
}
\description{
Process `quick_assoc()` outputs for easier visualization.
Perform p value adjustment, filter based on max p value
and exclude fixed loci.
}
