% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_association.R
\name{quick_assoc}
\alias{quick_assoc}
\title{Quick fisher test association function}
\usage{
quick_assoc(markers, pheno, test = "fisher")
}
\description{
Simple function to perform a fisher test for each
marker against a phenotype.
}
