% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_association.R
\name{run_assoc_logistic}
\alias{run_assoc_logistic}
\title{Run logisti regression on a set of CNV markers}
\usage{
run_assoc_logistic(
  markers,
  scan_res,
  pheno,
  padj = F,
  max_pval = 0.1,
  min_carriers = 5
)
}
\description{
Run logisti regression on a set of CNV markers
}
