% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_png_for_prediction.R
\name{save_pngs_prediction}
\alias{save_pngs_prediction}
\title{Save the PNG Plots of CNVs for Prediction}
\usage{
save_pngs_prediction(
  root,
  cnvs,
  samps,
  snps,
  w = 64,
  in_out_ratio = 3,
  shrink_lrr = 0.2
)
}
\arguments{
\item{root}{root folder for the dataset. Must not exists.}

\item{cnvs}{cnv data.table in the usual format}

\item{samps}{sample list in usual format}

\item{snps}{snps in the usual format}

\item{w}{see plot_cnv() must be the same the model used in training}

\item{in_out_ratio}{see load_snps_tbx(), must be the same the model used in training}

\item{shrink_lrr}{see load_snps_tbx(), should be the same the model used in training}
}
\description{
This function is used to create the dataset of PNG images for prediction
}
