% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCNVRs_IOU.R
\name{cnvrs_iou}
\alias{cnvrs_iou}
\title{Create CNVRs in a set of CNVS based on network analysis}
\usage{
cnvrs_iou(
  cnvs,
  chr_arm,
  window_size = 5e+05,
  min_iou = 0.75,
  leiden_res = 1,
  plots_path = NA,
  min_n = 10
)
}
\arguments{
\item{min_iou}{minimum IOU filter. Define how similar two CNVs must be in order to be considered connected}

\item{leiden_res}{lower value will tend to create more smaller communities}

\item{cnv}{usual CNVs `data.table`}

\item{chr_arms}{chromsome arms location, from `QCtreeCNV` package}

\item{plot_path}{not yet implemented}
}
\description{
A network is defined a set of CNVs that overlap with each other. Usually there
are several ones in each chromosomal arm. Given a minimum IOU value
each pairs of CNVs is considered as two connected nodes. The CNVRs are
computed using community the detection algorithm 'leiden'
}
