% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iou_matrix.R
\name{iou_matrix}
\alias{iou_matrix}
\title{Compute the Intersection Over the Union for a set of CNVs}
\usage{
iou_matrix(cnvs, chr_arms)
}
\arguments{
\item{cnv}{usual CNVs `data.table`}
}
\description{
Can be usefull for exploratory reasons but also as a base to
construct CNVRs.
}
