% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tabix.R
\name{load_snps_tbx}
\alias{load_snps_tbx}
\title{Load intensity data in R from tabix file}
\usage{
load_snps_tbx(
  cnv,
  samp,
  snps = NULL,
  in_out_ratio = 1,
  adjusted_lrr = T,
  min_lrr = -1.2,
  max_lrr = 1,
  shrink_lrr = NULL
)
}
\arguments{
\item{cnv}{one line data.table in the usual cnv format}

\item{samp}{one line samples file in the usual format}

\item{snps}{the snps file for PennCNV, as data.table}

\item{in_out_ratio}{ratio of bp outside the cnv vs inside, per side. A value of one means
the CNV length will be added on each side}

\item{adjusted_lrr}{load GC-adjusted LRR on normal LRR (T/F)}

\item{min_lrr}{minimum LRR value (lower values are set to the min_lrr)}

\item{max_lrr}{maximum LRR value (higher values are set to max_lrr)}

\item{shrink_lrr}{shrink LRR values toword the mean (done separately for
SNPs before / in / after the CNV}
}
\description{
A function to lead the snps data from the tabix indexed intensity file.
It will also perform some preprocessing, e.g. reduce the LRR interval to [-1.4 , 1.2],
}
