% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cnv_pixel_image.R
\name{plot_cnv}
\alias{plot_cnv}
\title{Create the matrix for the PNG image}
\usage{
plot_cnv(
  cnv,
  samp,
  tmp_plot = 0,
  shrink_lrr = 0.2,
  tabix_data = NULL,
  min_lrr = -1.4,
  max_lrr = 1.3,
  blank_small = F,
  w = 96,
  z = 4,
  k1 = 31,
  k2 = 26,
  l_wind = 2e+07,
  mx_lr = 2
)
}
\arguments{
\item{cnv}{see load_snps_tbx() documentation}

\item{samp}{see load_snps_tbx() documentation}

\item{tmp_plot}{for developing, if set to 1 plot the "normal"
LRR/BAF plot, if set to 2 plot the pixelated image in R}

\item{shrink_lrr}{shrink LRR values toword the mean (done separately for
SNPs before / in / after the CNV}

\item{min_lrr}{minimum LRR value (lower values are set to the min_lrr)}

\item{max_lrr}{maximum LRR value (higher values are set to max_lrr)}
}
\description{
This function create the pixel matrix that can be saved
as a PNG for further use. It will also perform some preprocessing, e.g.
reduce the LRR interval to [-1.4 , 1.2],
}
