% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cnv_pixel_image.R
\name{plot_cnv}
\alias{plot_cnv}
\title{Create the matrix for the PNG image}
\usage{
plot_cnv(
  cnv,
  samp,
  snps = NULL,
  in_out_ratio = 1,
  adjusted_lrr = T,
  w = 64,
  z_ratio = 0.1,
  tmp_plot = 0,
  min_lrr = -1.2,
  max_lrr = 1,
  shrink_lrr = NULL
)
}
\arguments{
\item{cnv}{see load_snps_tbx() documentation}

\item{samp}{see load_snps_tbx() documentation}

\item{snps}{see load_snps_tbx() documentation}

\item{in_out_ratio}{see load_snps_tbx() documentation}

\item{adjusted_lrr}{see load_snps_tbx() documentation}

\item{w}{size of square side in pixel}

\item{z_ratio}{approximate ratio of z (the blank section between
the two halves) compared to w}

\item{tmp_plot}{for developing, if set to 1 plot the "normal"
LRR/BAF plot, if set to 2 plot the pixelated image in R}

\item{min_lrr}{see load_snps_tbx() documentation}

\item{max_lrr}{see load_snps_tbx() documentation}

\item{shrink_lrr}{see load_snps_tbx() documentation}
}
\description{
This function create the pixel matrix that can be saved
as a PNG for further use. All processing is done in the
function load_snps_tbx().
}
