% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_png_training_dataset.R
\name{save_pngs_dataset}
\alias{save_pngs_dataset}
\title{Create Dataset of Validated CNVs}
\usage{
save_pngs_dataset(
  root,
  cnvs,
  samps,
  snps,
  shrink_lrr = 0.2,
  flip_chance = 0.5,
  hole_chance = 0.2,
  noise_chance = 0.15,
  noise_lvl = 0.1
)
}
\arguments{
\item{root}{root folder for the dataset. Must not exists.}

\item{cnvs}{cnv data.table in the usual format plus the column 'Visual_Output'}

\item{samps}{sample list in usual format}

\item{snps}{snps in the usual format}

\item{shrink_lrr}{see load_snps_tbx()}

\item{flip_chance}{probability of saving a flipped example as well}

\item{w}{see plot_cnv()}

\item{in_out_ratio}{see load_snps_tbx()}
}
\description{
This function is used to create the dataset of PNG images for training.
At the moment not all options for plot_cnv() are available here.
}
